Move config files to extra package
