+++ usecases/auth/authorization/rbac/rbacconf/config.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package rbac
+package rbacconf
 
 import "fmt"
