+++ usecases/auth/authorization/rbac/rbacconf/config_test.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package rbac
+package rbacconf
 
 import (
 	"testing"
